Move code for generating transition system into separate file
This moves the code for generating the transition system from
transition_system.cpp into build_transition_system.cpp.
