+++ src/ebmc/Makefile
@@ -1,6 +1,7 @@
 SRC = \
       bdd_engine.cpp \
       bmc.cpp \
+      build_transition_system.cpp \
       cegar/abstract.cpp \
       cegar/bmc_cegar.cpp \
       cegar/latch_ordering.cpp \
