+++ src/ebmc/transition_system.h
@@ -13,9 +13,6 @@ Author: Daniel Kroening, dkr@amazon.com
 #include <util/std_expr.h>
 #include <util/symbol_table.h>
 
-class cmdlinet;
-class message_handlert;
-
 class transition_systemt
 {
 public:
@@ -33,13 +30,5 @@ public:
   std::vector<symbol_exprt> state_variables() const;
   std::vector<symbol_exprt> inputs() const;
 };
-
-transition_systemt get_transition_system(const cmdlinet &, message_handlert &);
-
-int preprocess(const cmdlinet &, message_handlert &);
-int show_parse(const cmdlinet &, message_handlert &);
-int show_modules(const cmdlinet &, message_handlert &);
-int show_module_hierarchy(const cmdlinet &, message_handlert &);
-int show_symbol_table(const cmdlinet &, message_handlert &);
 
 #endif // CPROVER_EBMC_TRANSITION_SYSTEM_H
